spec: fix tempest tests directory
Tempest tests are in mistral_tempest_tests but the Tempest plugins is
looking at some files (Mistral Workflow resources) in mistral directory.

So we should follow how mistral directories are structured so we can run
Tempest.

This patch move:
%{python2_sitelib}/mistral_tempest_tests
to:
%{python2_sitelib}/mistral/mistral_tempest_tests
and also stop copying mistral_tempest_tests in %{python2_sitelib}. Tests
are failing, and it was not the way to go.

Change-Id: I66b1ac2c5bb891f7c95c0a5b2ffb5bef094d70f9