Use individual log files for each Mistral process
At the moment debugging and following mistral errors is harder because
the output is combined. Typically you want to find problems with your
workflow and actions in the engine log, the other two logs would generally
be errors in Mistral itself. Additionally it is hard to tell which of the
processes is emitting some of the log messages.
 Typically you either want to find errors in the
Change-Id: Ibb8545102d8f38c6b0c1982425be7744f0f22ab0