Update patch set 1
Patch Set 1: Code-Review+1

This is a great idea. Much better than what we have now.

Just to be clear tho - writing to disk like this does conflict with some Mistral design choices. Mistral is written to be distributed and if we ever have multiple underclouds running multiple instances of Mistral, the files may end up on one but not the other and things could then randomly fail as the execution is distributed.

Patch-set: 1
Reviewer: Gerrit User 180 <180@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1