Fix mistral home directory
Previously we used /home/mistral but this is never created for the
mistral user. This generally isn't a problem except when the commands
being run by mistral are using system lookups to determine the user's
home directory instead of honoring $HOME.  This change updates the
mistral home director to be /var/lib/mistral which exists and is
writable by this mistral user.

Change-Id: If13721347469f40d14b5e188d24ba8a06809e9ef
Related-Bug: LP#1778269