Create python2/python3 package depending on distro
This patch shows a way to make specs compatible with python2 for
el7 and python3 of el8/fedora in such a way that only one package
is created for each one.

A set of new macros is added that can be use to make compatibility
easier. These macros can be added in a macros packages instead of
in the spec if needed.

Depends-On: https://review.rdoproject.org/r/#/c/16331/

Change-Id: I718848ff542ad154f9c142b7a7b2596169f5e6e5