Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > re. EL7 to EL8 upgrades: this will not work from python2-* to
 > python3-* as far as RPM is concerned they are parallel installable.
 > Not sure what is the plan for that in https://leapp-to.github.io/upgrades


We must consider if we need/want to support package upgrades puth in RDO from EL7 to EL8. According to previous conversations, my understanding is that we are NOT supporting it as main deployments should be done using containers by then, and users not using containers will require to reinstall the system (I think nobody will be testing non-containerized upgrades from el7 to el8). If this is not the case and we are supporting in-place upgrades from el7 to el8 with RDO we need to consider it as a requirement and probably obsoleting python2-* with python3-* or something like that.

Patch-set: 2