Optimize systemd dependency for containers
According to https://fedoraproject.org/wiki/Packaging:Scriptlets:
If a package is suitable for installation without systemd (in a
container image, for example) and does not require any of the
systemd mechanisms such as tmpfiles.d, then the systemd_ordering macro
MAY be used instead of the systemd_requires macro.

That is exactly the case we want to address for container images
when installing packages in it.

Related-Bug: #1804822

Change-Id: If8b4b3b7bd056a7d38c43d07c9ce6da1350718e0
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>